src/model_deploy/requirements.txt
aws-cdk-lib
aws-cdk-constructs
yaml